pyproject.toml
[tool.pdm]

[tool.pdm.scripts]
dev = "python manage.py runserver"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
mi = "python manage.py migrate"
mm = "python manage.py makemigrations"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"

[tool.pdm.dev-dependencies]
dev = [
    "isort>=5.12.0",
    "black>=23.7.0",
]
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Paulo", email = "bruh@bruh.bruh"},
]
dependencies = [
    "django>=4.2.4",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=3.14.0",
    "djangorestframework-simplejwt>=5.2.2",
    "setuptools>=68.0.0",
    "python-magic>=0.4.27",
    "Pillow>=10.0.0",
    "django-extensions>=3.2.3",
    "pydotplus>=2.0.2",
    "drf-spectacular>=0.26.4",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"